A user can authenticate
This is very ugly right now. If the user provides a valid email/password
combination they will get a JWT.

The UI for everything is ugly. The form is almost illegible. It is
completely inaccessible, without labels of any kind. The error state is
not shown to the user in any way.

There is no content to put behind auth yet.

The JWT is not stored in localstorage.

There is no mechanism for logging out.